+++ pip_check_reqs/find_missing_reqs.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 import argparse
-import collections
 import logging
 import sys
 from pathlib import Path
@@ -35,51 +34,8 @@ def find_missing_reqs(
     )
     used_modules = imported_modules.found
 
-    installed_files: dict[Path, str] = {}
-    packages_info = common.get_packages_info()
-
-    for package in packages_info:
-        package_name = package.name
-        package_location = package.location
-
-        log.debug(
-            "installed package: %s (at %s)",
-            package_name,
-            package_location,
-        )
-        for item in package.files or []:
-            path = Path(package_location) / item
-            path = common.cached_resolve_path(path=path)
-
-            installed_files[path] = package_name
-            package_path = common.package_path(path=path)
-            if package_path:
-                # we've seen a package file so add the bare package directory
-                # to the installed list as well as we might want to look up
-                # a package by its directory path later
-                installed_files[package_path] = package_name
-
     # 3. match imported modules against those packages
-    used: collections.defaultdict[
-        NormalizedName,
-        list[common.FoundModule],
-    ] = collections.defaultdict(list)
-    for modname, info in used_modules.items():
-        # probably standard library if it's not in the files list
-        if info.filename in installed_files:
-            used_name = canonicalize_name(name=installed_files[info.filename])
-            log.debug(
-                "used module: %s (from package %s)",
-                modname,
-                installed_files[info.filename],
-            )
-            used[used_name].append(info)
-        else:
-            log.debug(
-                "used module: %s (from file %s, assuming stdlib or local)",
-                modname,
-                info.filename,
-            )
+    used = common.used_packages(used_modules=used_modules, paths=paths)
 
     # 4. compare with requirements
     explicit: set[NormalizedName] = set()